+++ Sources/EudiWalletKit/Services/PresentationService.swift
@@ -21,7 +21,8 @@ import MdocDataTransfer18013
 public typealias RequestItems = [String: [String: [String]]]
 
 /// Presentation service abstract protocol
-public protocol PresentationService {
+
+public protocol PresentationService: Sendable {
 	/// Status of the data transfer
 	//var status: TransferStatus { get }
 	/// instance of a presentation ``FlowType``
@@ -31,12 +32,12 @@ public protocol PresentationService {
 	///
 	/// - Returns: The requested items.
 	/// Receive request.
-	func receiveRequest() async throws -> [String: Any]
+	func receiveRequest() async throws -> UserRequestInfo
 	/// Send response to verifier
 	/// - Parameters:
 	///   - userAccepted: True if user accepted to send the response
 	///   - itemsToSend: The selected items to send organized in document types and namespaces (see ``RequestItems``)
-	func sendResponse(userAccepted: Bool, itemsToSend: RequestItems, onSuccess: ((URL?) -> Void)?) async throws
+	func sendResponse(userAccepted: Bool, itemsToSend: RequestItems, onSuccess: ( @Sendable (URL?) -> Void)?) async throws
 }